+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix store.getState is not a function error. (UIEH-1216)
 * Fix Detail Record: Cannot add a tag. (UIEH-1215)
+* Remove sticky list scroll on the details page. (UIEH-1214)
 
 ## [7.0.0] (https://github.com/folio-org/ui-eholdings/tree/v7.0.0) (2021-10-06)